ng2_apps/photos/src/app/save-dialog/save-dialog.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { SaveDialogComponent } from './save-dialog.component';

describe('Component: SaveDialog', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [SaveDialogComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([SaveDialogComponent],
      (component: SaveDialogComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(SaveDialogComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(SaveDialogComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-save-dialog></app-save-dialog>
  `,
  directives: [SaveDialogComponent]
})
class SaveDialogComponentTestController {
}

